backend/requirements.txt
# Backend API Requirements
# Battery RUL Prediction & Monitoring System

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# WebSocket Support
python-socketio==5.11.0
python-engineio==4.9.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# HTTP Client (for service-to-service)
httpx==0.26.0
tenacity==8.2.3

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Data Processing & Parquet
pandas==2.1.4
pyarrow==15.0.0
numpy==1.26.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0

